requirements.txt
Django==4.2.5
Pillow==10.0.0

django-allauth==0.56.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
django-extra-fields==3.0.2
drf-yasg==1.21.7

eventlet==0.33.3

stripe==6.6.0

psycopg2==2.9.7

redis==4.6.0
django-redis==5.3.0
celery==5.3.4


